FinTraka/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
whitenoise==6.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.3.0